requirements.txt
Flask==3.0.2
pandas==2.2.2
openpyxl==3.1.5
requests==2.32.3
jsonschema==4.23.0
python-dotenv==1.0.1
